moves SessionPersist to separate file
